Add update_manifest command, reorganize manifest functions
