+++ blaze/command/preprocess.py
@@ -3,9 +3,7 @@ from blaze.chrome.devtools import capture_har_in_mahimahi
 from blaze.config.client import get_default_client_environment
 from blaze.config.config import get_config
 from blaze.config.environment import EnvironmentConfig
-from blaze.evaluator.simulator import Simulator
 from blaze.logger import logger as log
-
 from blaze.preprocess.har import har_entries_to_resources
 from blaze.preprocess.record import record_webpage, find_url_stable_set
 from blaze.preprocess.resource import resource_list_to_push_groups
@@ -72,37 +70,3 @@ def preprocess(args):
     )
     env_config.save_file(args.output)
     log.info("successfully prepared website for training", output=args.output)
-
-
-@command.argument("manifest_file", help="The file path to the saved manifest file from `blaze preprocess`")
-@command.argument("--trainable", "-t", help="Only show trainable push groups", action="store_true", default=False)
-@command.argument("--verbose", "-v", help="Show more information", action="store_true", default=False)
-@command.command
-def view_manifest(args):
-    """ View the prepared manifest from `blaze preprocess` """
-    log.info("loading manifest", manifest_file=args.manifest_file)
-    env_config = EnvironmentConfig.load_file(args.manifest_file)
-
-    print("[[ Request URL ]]\n{}\n".format(env_config.request_url))
-    print("[[ Replay Dir ]]\n{}\n".format(env_config.replay_dir))
-    print("[[ Trainable Groups ]]\n{}\n".format("\n".join(group.name for group in env_config.trainable_push_groups)))
-    print("[[ {}Push Groups ]]".format("Trainable " if args.trainable else ""))
-
-    for group in env_config.push_groups:
-        if args.trainable and not group.trainable:
-            continue
-        print("  [{id}: {name} ({num} resources)]".format(id=group.id, name=group.name, num=len(group.resources)))
-        for res in group.resources:
-            url = Url.parse(res.url).resource
-            if len(url) > 64:
-                url = url[:61] + "..."
-            print(
-                "    {order:<3}  {url:<64}  {type:<6}  {size} B".format(
-                    order=res.order, url=url, type=res.type.name, size=res.size
-                )
-            )
-        print()
-
-    print("[[ Execution Graph ]]")
-    sim = Simulator(env_config)
-    sim.print_execution_map()
